Examples.Account.AbstractDomain

Depiction of the domain

type Payment = (String, Integer, String)

A payment contains a sender, an amount and a recipient

type Policy

Arguments

= Payment
The payment currently validated

-> Integer
The funds possessed by the current user

-> String
The current user for which validation is called

-> Bool
Returns whether the payment should go through

The type for policies

type Account = (Integer, Set String)

An account contains the balance and set of policies

data Register

The register associates:

policy names with policies

user names with accounts

Constructors

Register

Fields

policies :: Map String Policy

accounts :: Map String Account

class Monad m => MonadAccounts m where

Our domain specification: users can be added, payments can be attempted and balances can be requested.

Methods

addUser :: String -> Integer -> m ()

addPolicy :: String -> Policy -> m ()

allPolicies :: m [String]

subscribeToPolicy :: String -> String -> m ()

unsubscribeToPolicy :: String -> String -> m ()

simulate :: m a -> m (Maybe a)

issuePayment :: Payment -> m ()

getUserBalance :: String -> m Integer

data MonadAccountsEffect :: Effect where

Effects for the abstract domain

Constructors

AddUser :: forall (m :: Type -> Type). String -> Integer -> MonadAccountsEffect m ()

AddPolicy :: forall (m :: Type -> Type). String -> Policy -> MonadAccountsEffect m ()

AllPolicies :: forall (m :: Type -> Type). MonadAccountsEffect m [String]

SubscribeToPolicy :: forall (m :: Type -> Type). String -> String -> MonadAccountsEffect m ()

UnsubscribeToPolicy :: forall (m :: Type -> Type). String -> String -> MonadAccountsEffect m ()

Simulate :: forall (m :: Type -> Type) (a :: Type). (m a) -> MonadAccountsEffect m (Maybe a)

IssuePayment :: forall (m :: Type -> Type). Payment -> MonadAccountsEffect m ()

GetUserBalance :: forall (m :: Type -> Type). String -> MonadAccountsEffect m Integer