Examples.Account.SimpleDomain

Documentation

data AccountsError

Errors raised by the domain

Constructors

NoSuchAccount String

AlreadyExistingAccount String

NoSuchPolicy String

AlreadyExistingPolicy String

PolicyError

type AccountsT m = ExceptT AccountsError (StateT Register m)

Our domain implementation

Some helper functions over AccountsT m

ensureExistingUser :: Monad m => String -> AccountsT m Account

Ensures a user is registered and returns the associated account

ensureNonExistingUser :: Monad m => String -> AccountsT m ()

Ensures a user is not registered

modifyAccounts :: Monad m => (Map String Account -> Map String Account) -> AccountsT m ()

Modifies the current map of accounts

ensureExistingPolicy :: Monad m => String -> AccountsT m Policy

Ensures a policy is registered and executes an action based on its value

ensureNonExistingPolicy :: Monad m => String -> AccountsT m ()

Ensures a policy is not present and executes an action

modifyPolicies :: Monad m => (Map String Policy -> Map String Policy) -> AccountsT m ()

Modifies the current map of policies

runAccountsT :: Monad m => Register -> AccountsT m a -> m (Either AccountsError a, Map String Account)

A function to run the domain and return relevant information