Logic.Ltl

Description

Generate state-aware modifications of sequences of stateful actions. The approach is to use an LTL-inspired language to build composite modifications that apply atomic modifications to specific steps. Each atomic modification's applicability and parameters may depend on the state reached so far, and there will be zero or more possibilities to apply each composite modification.

The workflow is to

write effect types for all actions that you want to apply atomic modifications to,

write an abstract type describing your atomic modifications, together with a Semigroup instance that describes how they combine,

write instances of InterpretLtl (or InterpretLtlHigherOrder for higher-order effects) that explain how your atomic modifications apply to your effects,

use modifyLtl to apply composite modifications to your trace, and

use interpretLtlAST to run all modified versions of your trace.

The module Examples.Ltl.Simple contains a tutorial.

For historic context, this is a re-implementation of the LTL module from cooked-validators. The version this is based upon is: https://github.com/tweag/cooked-validators/blob/a460c1718d8d21bada510d83b705b24b0f8d36e0/src/Cooked/Ltl.hs

Ltl formulas

data Ltl a

Type of "LTL" formulas. Think of a as a type of atomic "modifications", then a value of type Ltl a describes a composite modification that describes where to apply these modifications.

Since it does not make (obvious) sense to talk of a negated modification or of one modification (possibly in the future) to imply another modification, implication and negation are absent.

Constructors

LtlTruth
The "do nothing" modification that is always applicable

LtlFalsity
The modification that never applies

LtlAtom a
The modification that applies a given atomic modification at the current step

LtlOr (Ltl a) (Ltl a)
Branch into the "timeline" where the left modification is applied and the one where the right modification is applied. (In a sense, this is an exclusive or, as we do not introduce the branch where both modifications are applied.)

LtlAnd (Ltl a) (Ltl a)
Apply both the left and the right modification. Attention: The "apply both" operation for atomic modifications of type a will be user-defined through a Semigroup instance. If that operation isn't commutative, this conjunction may also fail to be commutative.

LtlNext (Ltl a)
Apply the given modification at the next step.

LtlUntil (Ltl a) (Ltl a)
Apply the first modification at least until the second one begins to be applicable (and is applied), which must happen eventually. The formulas

a `LtlUntil` b

and

b `LtlOr` (a `LtlAnd` LtlNext (a `LtlUntil` b))

are equivalent.

LtlRelease (Ltl a) (Ltl a)
Apply the second modification up to and including the step when the first one becomes applicable (and is applied); if that never happens, the second formula will be applied forever. View this as a dual to LtlUntil. The formulas

a `LtlRelease` b

and

b `LtlAnd` (a `LtlOr` LtlNext (a `LtlRelease` b))

are equivalent.

somewhere :: a -> Ltl a

Apply an atomic modification to some action.

everywhere :: a -> Ltl a

Apply an atomic modification to all actions.

there :: Integer -> Ltl a -> Ltl a

Apply a modification from the n-th step in the trace (0-indexed).

Deploying Ltl formulas

type LtlAST mod ops = AST (LtlEffect mod ': ops)

An AST that allows modifying parts of its contents with Ltl modifications, using modifyLtl.

modifyLtl :: Ltl mod -> LtlAST mod ops a -> LtlAST mod ops a

Apply an Ltl modification to an LtlAST. Think of modifyLtl x acts as "try all ways to apply x to the actions given in acts".

Higher-order effects

class InterpretLtlHigherOrder (mod :: Type) (m :: Type -> Type) (op :: Effect) where

Explain how to interpret an Ltl modification, in the presence of higher-order effects. The type parameters have the same meaning as for InterpretLtl.

Methods

interpretLtlHigherOrder :: op (AST ops) a -> LtlInterpHigherOrder mod m ops a

Given an operation of type op a, there are two possibilities, corresponding the two constructors of LtlInterpHigherOrder.

For simple operations that don't "nest" other ASTs, use the Direct constructor. Its meaning corresponds precisely to the interpretLtl function.

For operations that do nest, use the Nested constructor. It needs some explanation: the stepwise approach based on applying atomic modifications to single operations breaks down for higher-order operations, since a single higher-order operation may contain ASTs of operations. We'll likely want to use a composite modification while evaluating such nested ASTs.

Composite modifications in the current setting are list of Ltl formulas. Each modifyLtl adds another formula to the head of that list, and all formulas are evaluated in parallel to the interpretation of the AST. (That is: if you don't nest modifyLtls, the list will only ever contain one element. If you nest modifyLtls, the head of the list will be the formula that was introduced by the innermost modifyLtl)

The Nested constructor proposes the following approach: It'll just give you a function

evalAST :: forall b. [Ltl mod] -> AST ops b -> m (b, [Ltl mod])

which you can call on the nested ASTs, which it'll evaluate with the list of Ltl formulas you provide. To explain it by example, let's use ErrorEffect:

instance (MonadError e m) => InterpretLtlHigherOrder x m (ErrorEffect e) where
  interpretLtlHigherOrder (CatchError acts handler) =
    Nested $ \evalAST ltls ->
      catchError
        (evalAST ltls acts)
        ( \err ->
            do
              (a, _) <- evalAST [] $ handler err
              return (a, ltls)
        )
  interpretLtlHigherOrder (ThrowError err) = ...

The equation for CatchError means that you'll interpret the body acts with the composite modification currently in place. If any error is thrown, you'll run the handler, without any modifications at all, and restore the original composite modification. There might be other ways to implement this nesting behaviour, depending on your use case, and the Nested constructor should hopefully be general enough to accommodate most of them.

data LtlInterpHigherOrder (mod :: Type) (m :: Type -> Type) (ops :: [Effect]) (a :: Type) where

codomain of interpretLtlHigherOrder. See the explanation there.

Constructors

Direct :: ModInterp mod m a -> LtlInterpHigherOrder mod m ops a

Nested :: ((forall b. [Ltl mod] -> AST ops b -> m (b, [Ltl mod])) -> [Ltl mod] -> m (a, [Ltl mod])) -> LtlInterpHigherOrder mod m ops a

Interpreting Ltl modifications

data LtlInstanceKind

Used to signify which instance is to be used for a specific effect in the InterpretEffectsLtl class.

Constructors

InterpretModTag

InterpretLtlHigherOrderTag

InterpretEffectStatefulTag

class InterpretEffectsLtl (mod :: Type) (m :: Type -> Type) (tags :: [LtlInstanceKind]) (ops :: [Effect]) where

The constraint used by interpretLtlASt and similar functions. See the documentation there.

Users of the library will never have to write instances of this class.

Methods

interpretEffectsLtl :: InterpretEffectsLtlConstraintList mod m tags ops

A witness of the constraint.

interpretLtlAST :: forall tags mod m ops a. (Semigroup mod, MonadPlus m, InterpretEffectsLtl mod m tags ops) => LtlAST mod ops a -> m a

Interpret an LtlAST into a suitable domain.

Each effect op in the list ops must have one of the following instances:

InterpretLtl mod m op

InterpretLtlHigherOrder mod m op

InterpretEffectStateful (Const [Ltl mod]) m op

Which instance is expected is declared through the tags. Since this type variable is ambiguous, you will have to type-apply this function to a list tags :: [LtlInstanceKind], which declare, in order, what instance you want to use for the operations in ops.

interpretLtlASTWithInitialFormulas :: forall tags mod m ops a. (Semigroup mod, MonadPlus m, InterpretEffectsLtl mod m tags ops) => [Ltl mod] -> LtlAST mod ops a -> m a

Like interpretLtlAST, just with an initial list of Ltl formulas that will be evaluated throughout the interpretation, even when there are no modifyLtls. Prunes all branches that end with incompletely applied Ltl formulas.

You'll also have to type-apply this to a list of LtlInstanceKinds, as described ad interpretLtlAST.

internal note: This function is the same as interpretASTStateful, just for an InterpretEffectsLtl constraint instead of InterpretEffectsStateful.